Added alot of changes
Added changes to styling and css for design improvements.

Added a read more and read less button, but its not working perfectly yet.

Added a button for the create new venue function to hide and declutter the profile page.

added a refresh site logic for the Create venue function to update the ''number of venues''